backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
pandas
matplotlib
seaborn
mlflow
python-multipart
python-dotenv